pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "A project to help patient get the best trial fitting them via AI Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mistralai>=1.10.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
